chore(download): remove migration-only characterization scaffolding (#781)
The Wave 1 golden-master tests existed to prove each refactor wave moved
behavior verbatim; that job is done now that the components have landed.

- Delete DownloadFilterCharacterizationTests and DownloadClassifierTests.
- Extract the shared TreeStubURLProtocol out of the deleted characterization
  file into its own helper (ProgressSequenceTests still depends on it).
- Drop stale doc-comment references to the deleted tests in ModelHub,
  HFTreeListerTests, and RetryPolicyTests.

Kept: the ModelHub download(...)/fetchFile(...) `configuration:` seams — they
remain load-bearing for the permanent DownloadResumeTests, HFTreeListerTests,
and FetchHuggingFaceFileTests.
